Fix build with CHERI_MIPS.
In particular, the gnulib bundled with gdb7.12 does a runtime check for
gettimeofday() that fails in cross-builds.  The replacement gettimeofday()
uses the 'restrict' keyword which is undefined in C++ (so clang chokes on
it).  The fix is to set an environment variable noting that gettimeofday
doesn't clobber (which FreeBSD's does not) to bypass the runtime test.